video-worker/requirements.txt
pika>=1.3.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
